backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
tinydb==4.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
openai
Pillow 
flask_session